Separate the config reader.
