+++ src/main/scala/com/kolor/docker/api/entities/ConfigReader.scala
@@ -1,0 +1,14 @@
+package com.kolor.docker.api.entities
+
+import scala.io.BufferedSource
+
+
+class ConfigReader(config: BufferedSource) {
+  private lazy val lines = config.getLines()
+
+  def getProperty(name: String): String =
+    lines
+      .find(_.startsWith(s"$name="))
+      .map(_.replace(s"$name=", ""))
+      .getOrElse("")
+}
